iox-#861 Rename typed_unique_id files to unique_port_id
