+++ iceoryx_posh/include/iceoryx_posh/mepoo/chunk_header.hpp
@@ -18,7 +18,7 @@
 #define IOX_POSH_MEPOO_CHUNK_HEADER_HPP
 
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
-#include "iceoryx_posh/internal/popo/building_blocks/typed_unique_id.hpp"
+#include "iceoryx_posh/internal/popo/building_blocks/unique_port_id.hpp"
 #include "iceoryx_posh/mepoo/chunk_settings.hpp"
 
 #include <cstdint>
